requirements.txt
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
Pillow==9.2.0
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.9.2
